pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "artifactory-mcp"
version = "0.1.0"
description = "MCP server for browsing and managing JFrog Artifactory artifacts."
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "dohq-artifactory>=1.0.1",
    "mcp[cli]>=1.26.0",
]

[project.scripts]
artifactory-mcp = "artifactory_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/artifactory_mcp"]

[tool.ruff]
lint.extend-select = ["I", "B", "UP", "SIM", "C4", "C90", "PIE", "RUF"]
lint.ignore = ["E501"]
lint.mccabe = { max-complexity = 14 }
lint.pydocstyle = { convention = 'google' }
line-length = 120
target-version = "py313"

[tool.mypy]
ignore_missing_imports = false
warn_unused_ignores = true
warn_redundant_casts = true
mypy_path = ["typings"]

[dependency-groups]
dev = [
    "prek>=0.3.3",
    "pytest>=9.0.2",
]
